+++ src/main-curriculum/components/custom-aside.tsx
@@ -1,11 +1,35 @@
-export const CustomAside = () => {
+import type { PersonalInfoResponse } from "../interfaces";
+import { getCandidateContactInfoMapped } from "../helpers";
+interface Props {
+  personalInfo: PersonalInfoResponse;
+}
+
+const BASE_URL = import.meta.env.VITE_BASE_URL;
+
+export const CustomAside = ({ personalInfo }: Props) => {
+  const canditateContactInfo = getCandidateContactInfoMapped(personalInfo);
+
   return (
-    <aside className="p-4">
-      Lorem ipsum dolor sit amet consectetur adipisicing elit. A laborum, culpa
-      nam nobis neque et nihil. Suscipit amet debitis modi architecto ea aliquam
-      placeat sed, inventore itaque veritatis exercitationem, quidem in quam
-      expedita! Animi ab explicabo, error velit omnis officia! Odit voluptas
-      totam minima ut ea sunt. Reprehenderit neque minus facilis tempore porro.
+    <aside className="bg-orange-50 p-4 flex flex-col items-center gap-6 h-full">
+      <h2 className="text-xl underline">Información Personal:</h2>
+      <img
+        className="w-2/4 md:w-1/2"
+        src={`${BASE_URL}/${personalInfo.photoUrl}`}
+        alt="foto del candidato"
+      />
+
+      <ul className="list-disc list-inside">
+        {Object.keys(canditateContactInfo).map((keyName: string) => {
+          const { key, value } = canditateContactInfo[keyName];
+          return (
+            <li key={keyName} className="py-1">
+              <strong>{key}:</strong>
+              <br />
+              <span>{value}</span>
+            </li>
+          );
+        })}
+      </ul>
     </aside>
   );
 };