+++ src/main-curriculum/components/custom-footer.tsx
@@ -1,3 +1,4 @@
+import { getCurrentYear } from "../helpers";
 interface Props {
   name: string;
 }
@@ -5,7 +6,7 @@ interface Props {
 export const CustomFooter = ({ name }: Props) => {
   return (
     <footer className="p-4 text-center text-white bg-blue-300 border-t-2 border-t-gray-50">
-      &copy; {name} (2025)
+      &copy; {name} ({getCurrentYear()})
     </footer>
   );
 };